requirements.txt
fastapi
uvicorn
opencv-python
numpy
pillow
python-multipart
matplotlib
